+++ src/api/v1/routes/invoice/invoice.route.ts
@@ -10,17 +10,25 @@ import { getInvoiceContract } from '../../contracts/invoice/get-invoice.contract
 export interface InvoiceRouteControllers {
   createInvoiceController: RequestHandler
   getInvoiceController: RequestHandler
+  createInvoiceMerchantRateLimitMiddleware?: RequestHandler
+  createInvoiceIpRateLimitMiddleware?: RequestHandler
+  getInvoiceIpRateLimitMiddleware?: RequestHandler
 }
 
 export const createInvoiceRoute = ({
   createInvoiceController,
   getInvoiceController,
+  createInvoiceMerchantRateLimitMiddleware = passThroughMiddleware,
+  createInvoiceIpRateLimitMiddleware = passThroughMiddleware,
+  getInvoiceIpRateLimitMiddleware = passThroughMiddleware,
 }: InvoiceRouteControllers): Router => {
   const router = Router()
 
   router.post(
     '/invoice',
     createContractBodyValidator(createInvoiceContract),
+    createInvoiceMerchantRateLimitMiddleware,
+    createInvoiceIpRateLimitMiddleware,
     createContractResponseValidator(createInvoiceContract),
     createInvoiceController,
   )
@@ -28,9 +36,14 @@ export const createInvoiceRoute = ({
   router.get(
     '/invoice/:id',
     createContractParamsValidator(getInvoiceContract),
+    getInvoiceIpRateLimitMiddleware,
     createContractResponseValidator(getInvoiceContract),
     getInvoiceController,
   )
 
   return router
+}
+
+const passThroughMiddleware: RequestHandler = (_req, _res, next) => {
+  return next()
 }